congress instance location thumbnails

+++ src/workers/http/routing/congresses/$congressId/instances/$instanceId/locations/$locationId/delete.js
@@ -1,6 +1,6 @@
-import path from 'path';
+import { deleteObjects } from 'akso/lib/s3';
 
-import { removePathAndEmptyParents } from 'akso/lib/file-util';
+import { thumbnailSizes } from './thumbnail/schema';
 
 export default {
 	schema: {
@@ -20,27 +20,29 @@ export default {
 		if (!orgData) { return res.sendStatus(404); }
 		if (!req.hasPermission('congress_instances.update.' + orgData.org)) { return res.sendStatus(403); }
 
-		const deleted = await AKSO.db('congresses_instances_locations')
+		const locData = await AKSO.db('congresses_instances_locations')
+			.where({
+				congressInstanceId: req.params.instanceId,
+				id: req.params.locationId
+			})
+			.first('thumbnailS3Id');
+		if (!locData) { return res.sendStatus(404); }
+
+		// Delete the thumbnail if it exists
+		if (locData.thumbnailS3Id) {
+			await deleteObjects({
+				keys: thumbnailSizes.map(size => `congresses-locations-thumbnails-${locData.thumbnailS3Id}-${size}`),
+			});
+		}
+
+		// Delete the location
+		await AKSO.db('congresses_instances_locations')
 			.where({
 				congressInstanceId: req.params.instanceId,
 				id: req.params.locationId
 			})
 			.delete();
 
-		if (deleted) {
-			const locPar = path.join(
-				AKSO.conf.dataDir,
-				'congress_instance_location_thumbnails',
-				req.params.congressId
-			);
-			const locPath = path.join(
-				locPar,
-				req.params.instanceId,
-				req.params.locationId
-			);
-			await removePathAndEmptyParents(locPar, locPath);
-		}
-
-		res.sendStatus(deleted ? 204 : 404);
+		res.sendStatus(204);
 	}
 };
